Refactored workhorse methods into libraries/default.rb
